# backend/user-backend-app/src/main/java/com/youssefdev/user/dto/FileHistoryResponse.java
package com.youssefdev.user.dto;

import java.time.Instant;

public record FileHistoryResponse(
        String action,
        String storedFileName,
        String originalFileName,
        Instant occurredAt
) {
}