# backend/user-backend-app/src/main/java/com/youssefdev/user/dto/FileInfoResponse.java
package com.youssefdev.user.dto;

import java.time.Instant;

public record FileInfoResponse(
        String storedFileName,
        String originalFileName,
        String contentType,
        long size,
        Instant createdAt,
        boolean passwordProtected
) {
}