# backend/user-backend-app/src/main/java/com/youssefdev/user/dto/RegisterRequest.java
package com.youssefdev.user.dto;

import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;

public record RegisterRequest(
        @NotBlank @Size(min = 3, max = 64) String login,
        @NotBlank @Size(min = 8, max = 128) String password
) {
}